refactor(telemetry): rename otel package to telemetry
The local otel package shadowed upstream go.opentelemetry.io/otel (forcing
otelSDK/feotel aliases). Renamed to telemetry; it will also host the bootstrap
(telemetry.Init, L3). Pure rename, no behaviour change.
